Moved readme out into main folder and updated formatting
Instead of having a repository with several Arduino projects, I'm going
back to the idea of one folder per project, as it seems to be the Git
way of doing things.
I'm keeping LiquidCrystalEAX in a sub-folder though so that it's easy to
copy into an Arduino installation.
I also moved the readme back up to the top-level folder, and formatted
it properly using Markdown.